refactor: rename Fieldset to LabelOutlined
